// packages/python/src/types.d.ts
export interface SimpleStatements {
    readonly $type: TSKindId.SimpleStatements;
    readonly $children: NonEmptyArray<SimpleStatement>;
}
export interface Suite {
    readonly $type: "_suite";
    readonly $children: readonly [SimpleStatements | Block | Newline];
}
// ...
export interface AsPattern {
    readonly $type: TSKindId.AsPattern;
    readonly _expression: Expression;
    readonly _alias: AsPatternTarget;
    expression(): Expression;
    alias(): AsPatternTarget;
}
// ...
export interface Attribute {
    readonly $type: TSKindId.Attribute;
    readonly _object: PrimaryExpression;
    readonly _attribute: Identifier;
    object(): PrimaryExpression;
    attribute(): Identifier;
}
// ...
export interface Block {
    readonly $type: TSKindId.Block;
    readonly $children: readonly (Statement)[];
}
export interface BooleanOperator {
    readonly $type: TSKindId.BooleanOperator;
    readonly _left: Expression;
    readonly _operator: AutoStamp<BooleanOperatorOperator>;
    readonly _right: Expression;
    left(): Expression;
    operator(): AutoStamp<BooleanOperatorOperator>;
    right(): Expression;
}
// ...
export interface ComparisonOperator {
    readonly $type: TSKindId.ComparisonOperator;
    readonly _left: PrimaryExpression;
    readonly _operators: NonEmptyArray<Bitflag<Operators, "<" | "<=" | "==" | "!=" | ">=" | ">" | "<>" | "in" | "not in" | "is" | "is not">>;
    left(): PrimaryExpression;
    operators(): NonEmptyArray<Bitflag<Operators, "<" | "<=" | "==" | "!=" | ">=" | ">" | "<>" | "in" | "not in" | "is" | "is not">>;
    readonly $children: NonEmptyArray<PrimaryExpression>;
}
// ...
export interface ConditionalExpression {
    readonly $type: TSKindId.ConditionalExpression;
    readonly _body: Expression;
    readonly _condition: Expression;
    readonly _alternative: Expression;
    body(): Expression;
    condition(): Expression;
    alternative(): Expression;
}
// ...
export interface ElseClause {
    readonly $type: TSKindId.ElseClause;
    readonly _body: Suite;
    body(): Suite;
}
// ...
export interface ExpressionList {
    readonly $type: TSKindId.ExpressionList;
    readonly $children: NonEmptyArray<Expression | "," | Expression>;
}
// ...
export interface ForStatement {
    readonly $type: TSKindId.ForStatement;
    readonly _async_marker?: BooleanKeyword<AsyncMarker>;
    readonly _left: LeftHandSide;
    readonly _right: Expressions;
    readonly _body: Suite;
    readonly _alternative?: ElseClause;
    asyncMarker(): BooleanKeyword<AsyncMarker> | undefined;
    left(): LeftHandSide;
    right(): Expressions;
    body(): Suite;
    alternative(): ElseClause | undefined;
}
// ...
export interface IfStatement {
    readonly $type: TSKindId.IfStatement;
    readonly _condition: Expression;
    readonly _consequence: Suite;
    readonly _alternative?: readonly (ElifClause | ElseClause)[];
    condition(): Expression;
    consequence(): Suite;
    alternative(): readonly (ElifClause | ElseClause)[];
}
// ...
export interface Lambda {
    readonly $type: TSKindId.Lambda;
    readonly _parameters?: LambdaParameters;
    readonly _body: Expression;
    parameters(): LambdaParameters | undefined;
    body(): Expression;
}
// ...
export interface ListPattern {
    readonly $type: TSKindId.ListPattern;
    readonly $children: readonly (Pattern)[];
}
export interface ListSplat {
    readonly $type: TSKindId.ListSplat;
    readonly _expression: Expression;
    expression(): Expression;
}
export interface ListSplatPattern {
    readonly $type: TSKindId.ListSplatPattern;
    readonly $children: readonly [Identifier | KeywordIdentifier | Subscript | Attribute];
}
// ...
export interface NamedExpression {
    readonly $type: TSKindId.NamedExpression;
    readonly _name: NamedExpressionLhs;
    readonly _value: Expression;
    name(): NamedExpressionLhs;
    value(): Expression;
}
export interface NonlocalStatement {
    readonly $type: TSKindId.NonlocalStatement;
    readonly $children: NonEmptyArray<Identifier>;
}
export interface NotOperator {
    readonly $type: TSKindId.NotOperator;
    readonly _argument: Expression;
    argument(): Expression;
}
// ...
export interface PatternList {
    readonly $type: TSKindId.PatternList;
    readonly $children: NonEmptyArray<Pattern | "," | Pattern>;
}
// ...
export interface Subscript {
    readonly $type: TSKindId.Subscript;
    readonly _value: PrimaryExpression;
    readonly _subscript: NonEmptyArray<Expression | Slice>;
    value(): PrimaryExpression;
    subscript(): NonEmptyArray<Expression | Slice>;
}
// ...
export interface TuplePattern {
    readonly $type: TSKindId.TuplePattern;
    readonly $children: readonly (Pattern)[];
}
// ...
export type AsyncMarker = Terminal<TSKindId.Async, "async">;
// ...
export type Identifier = Terminal<TSKindId.Identifier, string>;
// ...
export type Newline = Terminal<TSKindId.Newline, string>;
// ...
export type Expressions = ExpressionList;
export type ExpressionsTree = ExpressionListTree;
export type FExpression = ExpressionList | PatternList | Yield;
export type FExpressionTree = ExpressionListTree | PatternListTree | YieldTree;
export type LeftHandSide = PatternList;
// ...
export type Statement = SimpleStatements | IfStatement | ForStatement | WhileStatement | TryStatement | WithStatement | FunctionDefinition | ClassDefinition | DecoratedDefinition | MatchStatement;
export type StatementTree = SimpleStatementsTree | IfStatementTree | ForStatementTree | WhileStatementTree | TryStatementTree | WithStatementTree | FunctionDefinitionTree | ClassDefinitionTree | DecoratedDefinitionTree | MatchStatementTree;
export type Expression = ComparisonOperator | NotOperator | BooleanOperator | Lambda | ConditionalExpression | NamedExpression | AsPattern;
export type ExpressionTree = ComparisonOperatorTree | NotOperatorTree | BooleanOperatorTree | LambdaTree | ConditionalExpressionTree | NamedExpressionTree | AsPatternTree;
export type KeywordIdentifier = Identifier;
export type KeywordIdentifierTree = IdentifierTree;
export type Parameter = Identifier | TypedParameter | DefaultParameter | TypedDefaultParameter | ListSplatPattern | TuplePattern | DictionarySplatPattern;
export type ParameterTree = IdentifierTree | TypedParameterTree | DefaultParameterTree | TypedDefaultParameterTree | ListSplatPatternTree | TuplePatternTree | DictionarySplatPatternTree;
export type Pattern = Identifier | KeywordIdentifier | Subscript | Attribute | ListSplatPattern | TuplePattern | ListPattern;
// ...
export type ConfigFor<K extends keyof NamespaceMap> = NamespaceMap[K]['Config'];
export type FluentFor<K extends keyof NamespaceMap> = NamespaceMap[K]['Fluent'];
export type LooseFor<K extends keyof NamespaceMap> = NamespaceMap[K]['Loose'];
export type TreeFor<K extends keyof NamespaceMap> = NamespaceMap[K]['Tree'];
// ...
    type Config = ConfigFor<'_as_pattern'>;
    type Fluent = FluentFor<'_as_pattern'>;
    type Loose = LooseFor<'_as_pattern'>;
    type Tree = TreeFor<'_as_pattern'>;
    type Kind = '_as_pattern';
// ...
export declare namespace SimpleStatements {
    type Config = ConfigFor<'_simple_statements'>;
    type Fluent = FluentFor<'_simple_statements'>;
    type Loose = LooseFor<'_simple_statements'>;
    type Tree = TreeFor<'_simple_statements'>;
    type Kind = '_simple_statements';
}
export declare namespace Suite {
    type Config = ConfigFor<'_suite'>;
    type Fluent = FluentFor<'_suite'>;
    type Loose = LooseFor<'_suite'>;
    type Tree = TreeFor<'_suite'>;
    type Kind = '_suite';
}
// ...
export declare namespace AsPattern {
    type Config = ConfigFor<'as_pattern'>;
    type Fluent = FluentFor<'as_pattern'>;
    type Loose = LooseFor<'as_pattern'>;
    type Tree = TreeFor<'as_pattern'>;
    type Kind = 'as_pattern';
}
// ...
export declare namespace Attribute {
    type Config = ConfigFor<'attribute'>;
    type Fluent = FluentFor<'attribute'>;
    type Loose = LooseFor<'attribute'>;
    type Tree = TreeFor<'attribute'>;
    type Kind = 'attribute';
}
// ...
export declare namespace Block {
    type Config = ConfigFor<'block'>;
    type Fluent = FluentFor<'block'>;
    type Loose = LooseFor<'block'>;
    type Tree = TreeFor<'block'>;
    type Kind = 'block';
}
export declare namespace BooleanOperator {
    type Config = ConfigFor<'boolean_operator'>;
    type Fluent = FluentFor<'boolean_operator'>;
    type Loose = LooseFor<'boolean_operator'>;
    type Tree = TreeFor<'boolean_operator'>;
    type Kind = 'boolean_operator';
}
// ...
export declare namespace ComparisonOperator {
    type Config = ConfigFor<'comparison_operator'>;
    type Fluent = FluentFor<'comparison_operator'>;
    type Loose = LooseFor<'comparison_operator'>;
    type Tree = TreeFor<'comparison_operator'>;
    type Kind = 'comparison_operator';
}
// ...
export declare namespace ConditionalExpression {
    type Config = ConfigFor<'conditional_expression'>;
    type Fluent = FluentFor<'conditional_expression'>;
    type Loose = LooseFor<'conditional_expression'>;
    type Tree = TreeFor<'conditional_expression'>;
    type Kind = 'conditional_expression';
}
// ...
export declare namespace ElseClause {
    type Config = ConfigFor<'else_clause'>;
    type Fluent = FluentFor<'else_clause'>;
    type Loose = LooseFor<'else_clause'>;
    type Tree = TreeFor<'else_clause'>;
    type Kind = 'else_clause';
}
// ...
export declare namespace ExpressionList {
    type Config = ConfigFor<'expression_list'>;
    type Fluent = FluentFor<'expression_list'>;
    type Loose = LooseFor<'expression_list'>;
    type Tree = TreeFor<'expression_list'>;
    type Kind = 'expression_list';
}
// ...
export declare namespace ForStatement {
    type Config = ConfigFor<'for_statement'>;
    type Fluent = FluentFor<'for_statement'>;
    type Loose = LooseFor<'for_statement'>;
    type Tree = TreeFor<'for_statement'>;
    type Kind = 'for_statement';
}
// ...
export declare namespace Lambda {
    type Config = ConfigFor<'lambda'>;
    type Fluent = FluentFor<'lambda'>;
    type Loose = LooseFor<'lambda'>;
    type Tree = TreeFor<'lambda'>;
    type Kind = 'lambda';
}
// ...
export declare namespace ListPattern {
    type Config = ConfigFor<'list_pattern'>;
    type Fluent = FluentFor<'list_pattern'>;
    type Loose = LooseFor<'list_pattern'>;
    type Tree = TreeFor<'list_pattern'>;
    type Kind = 'list_pattern';
}
export declare namespace ListSplat {
    type Config = ConfigFor<'list_splat'>;
    type Fluent = FluentFor<'list_splat'>;
    type Loose = LooseFor<'list_splat'>;
    type Tree = TreeFor<'list_splat'>;
    type Kind = 'list_splat';
}
export declare namespace ListSplatPattern {
    type Config = ConfigFor<'list_splat_pattern'>;
    type Fluent = FluentFor<'list_splat_pattern'>;
    type Loose = LooseFor<'list_splat_pattern'>;
    type Tree = TreeFor<'list_splat_pattern'>;
    type Kind = 'list_splat_pattern';
}
// ...
export declare namespace NamedExpression {
    type Config = ConfigFor<'named_expression'>;
    type Fluent = FluentFor<'named_expression'>;
    type Loose = LooseFor<'named_expression'>;
    type Tree = TreeFor<'named_expression'>;
    type Kind = 'named_expression';
}
export declare namespace NonlocalStatement {
    type Config = ConfigFor<'nonlocal_statement'>;
    type Fluent = FluentFor<'nonlocal_statement'>;
    type Loose = LooseFor<'nonlocal_statement'>;
    type Tree = TreeFor<'nonlocal_statement'>;
    type Kind = 'nonlocal_statement';
}
export declare namespace NotOperator {
    type Config = ConfigFor<'not_operator'>;
    type Fluent = FluentFor<'not_operator'>;
    type Loose = LooseFor<'not_operator'>;
    type Tree = TreeFor<'not_operator'>;
    type Kind = 'not_operator';
}
// ...
export declare namespace PatternList {
    type Config = ConfigFor<'pattern_list'>;
    type Fluent = FluentFor<'pattern_list'>;
    type Loose = LooseFor<'pattern_list'>;
    type Tree = TreeFor<'pattern_list'>;
    type Kind = 'pattern_list';
}
// ...
export declare namespace Subscript {
    type Config = ConfigFor<'subscript'>;
    type Fluent = FluentFor<'subscript'>;
    type Loose = LooseFor<'subscript'>;
    type Tree = TreeFor<'subscript'>;
    type Kind = 'subscript';
}
// ...
export declare namespace TuplePattern {
    type Config = ConfigFor<'tuple_pattern'>;
    type Fluent = FluentFor<'tuple_pattern'>;
    type Loose = LooseFor<'tuple_pattern'>;
    type Tree = TreeFor<'tuple_pattern'>;
    type Kind = 'tuple_pattern';
}
// ...
    interface Transport {
        readonly $type: TSKindId._AsPattern;
        readonly $source?: 0 | 1 | 2;
        readonly $named?: boolean;
        readonly $text?: string;
        readonly $span?: {
            readonly start: number;
            readonly end: number;
        };
        readonly $nodeHandle?: number;
        readonly $childIndex?: number;
        readonly $children: readonly [CasePattern.Transport | Identifier.Transport];
    }
// ...
export declare namespace AsyncMarker {
    type Transport = boolean;
}
// ...
export declare namespace Identifier {
    type Transport = TerminalTransport<TSKindId.Identifier, string>;
}
// ...
export declare namespace Expressions {
    type Transport = ExpressionList.Transport;
}
export declare namespace FExpression {
    type Transport = ExpressionList.Transport | PatternList.Transport | Yield.Transport;
}
export declare namespace LeftHandSide {
    type Transport = PatternList.Transport;
}
// ...
export declare namespace Statement {
    type Transport = SimpleStatements.Transport | IfStatement.Transport | ForStatement.Transport | WhileStatement.Transport | TryStatement.Transport | WithStatement.Transport | FunctionDefinition.Transport | ClassDefinition.Transport | DecoratedDefinition.Transport | MatchStatement.Transport;
}
export declare namespace Expression {
    type Transport = ComparisonOperator.Transport | NotOperator.Transport | BooleanOperator.Transport | Lambda.Transport | ConditionalExpression.Transport | NamedExpression.Transport | AsPattern.Transport;
}
export declare namespace KeywordIdentifier {
    type Transport = Identifier.Transport;
}
export declare namespace Parameter {
    type Transport = Identifier.Transport | TypedParameter.Transport | DefaultParameter.Transport | TypedDefaultParameter.Transport | ListSplatPattern.Transport | TuplePattern.Transport | KeywordSeparator.Transport | PositionalSeparator.Transport | DictionarySplatPattern.Transport;
}
export declare namespace Pattern {
    type Transport = Identifier.Transport | Subscript.Transport | Attribute.Transport | ListSplatPattern.Transport | TuplePattern.Transport | ListPattern.Transport;
}